requirements/local.txt
asgiref==3.5.0
Django==4.0.1
django-model-utils==4.2.0
djangorestframework==3.13.1
Pillow==9.0.0
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
Unipath==1.1
